fix(searchbar): iconContext needs to be inside its own folder
