+++ src/drive/web/modules/services/components/helpers.js
@@ -1,46 +1,8 @@
 import { models } from 'cozy-client'
 
-import { getFileMimetype } from 'drive/lib/getFileMimetype'
-import { hasEncryptionRef } from 'drive/lib/encryption'
-
 export const TYPE_DIRECTORY = 'directory'
 
-const iconsContext =
-  typeof require.context === 'undefined' // no require.context in jest
-    ? { keys: () => [] }
-    : require.context('drive/assets/icons/', false, /icon-type-.*.svg$/)
-
-const icons = iconsContext.keys().reduce((acc, item) => {
-  acc[item.replace(/\.\/icon-type-(.*)\.svg/, '$1')] = iconsContext(item)
-  return acc
-}, {})
-
-const getIconUrl = file => {
-  let keyIcon
-  if (file.type === TYPE_DIRECTORY) {
-    if (hasEncryptionRef(file)) {
-      keyIcon = 'encrypted-folder'
-    } else {
-      keyIcon = 'folder'
-    }
-  } else {
-    keyIcon = models.file.isNote(file)
-      ? 'note'
-      : getFileMimetype(icons)(file.mime, file.name) || 'files'
-  }
-  const icon = icons[keyIcon].default
-
-  return `data:image/svg+xml;utf8,<svg xmlns='http://www.w3.org/2000/svg' width='32' height='32' viewBox='${icon.viewBox}'>${icon.content}<use href='#${icon.id}' x='0' y='0' width='32' height='32'/></svg>`.replace(
-    /#/g,
-    '%23'
-  )
-}
-
-export const containerForTesting = {
-  getIconUrl
-}
-
-export const makeNormalizedFile = async (client, folders, file) => {
+export const makeNormalizedFile = async (client, folders, file, getIconUrl) => {
   const isDir = file.type === TYPE_DIRECTORY
   const dirId = isDir ? file._id : file.dir_id
   const urlToFolder = `${window.location.origin}/#/folder/${dirId}`
@@ -64,6 +26,6 @@ export const makeNormalizedFile = async (client, folders, file) => {
     name: file.name,
     path,
     url,
-    icon: containerForTesting.getIconUrl(file)
+    icon: getIconUrl(file)
   }
 }
